pyproject.toml
[tool.poetry]
    name        = "bell-avr-vmc-apriltag-module"
    version     = "0.1.0"
    description = "AprilTag detection module for Bell AVR"
    license     = "MIT"
    readme      = "README.md"
    homepage    = "https://github.com/bellflight/AVR-VMC-AprilTag-Module"
    authors     = ["Nathan Vaughn <nvaughn51@gmail.com>"]

[tool.poetry.dependencies]
    python             = ">=3.9,<3.12"
    loguru             = "0.7.0"
    bell-avr-libraries = "0.2.0a2"
    numpy              = "1.25.0"
    transforms3d       = "0.4.1"
    nptyping           = "2.5.0"

[tool.poetry.group.dev.dependencies]
    pre-commit      = "3.3.2"
    clang-format    = "16.0.6"
    pytest          = "7.4.0"
    pytest-mock     = "3.11.1"
    pytest-cov      = "4.1.0"
    opencv-python   = "4.8.0.74"     # for type-hinting
    pupil-apriltags = "1.0.4.post10" # for type-hinting

[tool.pyright]
    typeCheckingMode           = "basic"
    venvPath                   = "."
    venv                       = ".venv"
    reportMissingParameterType = true

[tool.pytest.ini_options]
    # some of the C libraries include their own Python tests
    # that we are not interested in
    addopts = "--ignore=src/c"
    # for nptyping
    filterwarnings = "ignore::DeprecationWarning"

[tool.ruff]
    ignore = ["E501"]
